requirements.txt
bs4~=0.0.2
fastapi~=0.115.12
redis
types-redis
uvicorn
pymongo~=4.13.0
requests~=2.32.3
beautifulsoup4~=4.13.4
pydantic~=2.11.5
python-dotenv~=1.1.0